Move WRITABLE_CONFIG_FIELDS + MODEL_ROLE_FIELDS to config_meta to drop the cli→server handlers dep
